修复缓存超时时间问题

+++ framework-cache/framework-cache-core/src/main/java/com/hbasesoft/framework/cache/core/annotation/Cache.java
@@ -30,5 +30,5 @@ public @interface Cache {
 
     String key() default GlobalConstants.BLANK;
 
-    long expireTime() default 0L;
+    int expireTime() default 0;
 }
